fix(mimo-v2.5): pass dtype to make_array_from_callback for non-addressable shards
The embed stage uses a single-CPU-device mesh; on non-rank0 hosts a shard can have
no addressable data, and jax.make_array_from_callback then cannot infer dtype from
its (uninvoked) callback -> ValueError. Pass dtype=target_dtype explicitly (the
known safetensors load dtype); the following .astype is then a no-op on the source.

+++ python/sgl_jax/srt/utils/weight_utils.py
@@ -1140,9 +1140,13 @@ class WeightLoader:
 
                 return _load_slice
 
-            lazy_array = jax.make_array_from_callback(shape, sharding, _make_load_slice()).astype(
-                target_dtype
-            )
+            # Pass dtype explicitly: when this host has no addressable shard for the
+            # given sharding (e.g. the embed stage's single-CPU-device mesh on non-rank0
+            # hosts), jax.make_array_from_callback cannot infer dtype from the callback
+            # (which is never invoked here) and raises. target_dtype is the load dtype.
+            lazy_array = jax.make_array_from_callback(
+                shape, sharding, _make_load_slice(), dtype=target_dtype
+            ).astype(target_dtype)
 
             lazy_arrays.append(lazy_array)
